docs(events): add javadocs to EventSystem stuffs
